staging

+++ src/main/java/rencanakan/id/talentpool/controller/ExperienceController.java
@@ -3,13 +3,13 @@ package rencanakan.id.talentpool.controller;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import rencanakan.id.talentpool.dto.ExperienceListResponseDTO;
 import rencanakan.id.talentpool.dto.ExperienceRequestDTO;
 import rencanakan.id.talentpool.dto.ExperienceResponseDTO;
 import rencanakan.id.talentpool.dto.WebResponse;
+import rencanakan.id.talentpool.model.User;
 import rencanakan.id.talentpool.service.ExperienceService;
-import rencanakan.id.talentpool.service.ExperienceServiceImpl;
 
 import java.util.List;
 
@@ -17,28 +17,50 @@ import java.util.List;
 @RequestMapping("/experiences")
 public class ExperienceController {
 
-    private final ExperienceServiceImpl experienceService;
+    private final ExperienceService experienceService;
 
-    public ExperienceController(ExperienceServiceImpl experienceService) {
+    public ExperienceController(ExperienceService experienceService) {
         this.experienceService = experienceService;
     }
 
-    @GetMapping("/{talent_id}")
+    @GetMapping("/user/{talent_id}")
     public ResponseEntity<WebResponse<List<ExperienceResponseDTO>>> getExperiencesByTalentId(
             @PathVariable("talent_id") String talentId,
-            @RequestHeader("Authorization") String token) {
+            @AuthenticationPrincipal User user) {
 
-        List<ExperienceResponseDTO> resp = experienceService.getByTalentId(talentId);
-        return ResponseEntity.ok(WebResponse.<List<ExperienceResponseDTO>>builder()
-                .data(resp)
-                .build());
+        System.out.println("masuk sini kok");
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(WebResponse.<List<ExperienceResponseDTO>>builder()
+                            .errors("Unauthorized access")
+                            .build());
+        }
+
+        try {
+            List<ExperienceResponseDTO> resp = experienceService.getByTalentId(talentId);
+            return ResponseEntity.ok(WebResponse.<List<ExperienceResponseDTO>>builder()
+                    .data(resp)
+                    .build());
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(WebResponse.<List<ExperienceResponseDTO>>builder()
+                            .errors(e.getMessage())
+                            .build());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(WebResponse.<List<ExperienceResponseDTO>>builder()
+                            .errors(e.getMessage())
+                            .build());
+        }
     }
 
     @PostMapping
     public ResponseEntity<WebResponse<ExperienceResponseDTO>> createExperience(
-            @RequestBody @Valid ExperienceRequestDTO request) {
-
-        ExperienceResponseDTO createdExperience = experienceService.createExperience(request);
+            @RequestBody @Valid ExperienceRequestDTO request,
+            @AuthenticationPrincipal User user
+    ) {
+        ExperienceResponseDTO createdExperience = experienceService.createExperience(user.getId(), request);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(WebResponse.<ExperienceResponseDTO>builder()
                         .data(createdExperience)
@@ -47,11 +69,11 @@ public class ExperienceController {
 
     @PutMapping("/{id}")
     public ResponseEntity<WebResponse<ExperienceResponseDTO>> editExperienceById(
+            @AuthenticationPrincipal User user,
             @PathVariable Long id,
-            @RequestHeader("Authorization") String token,
             @RequestBody @Valid ExperienceRequestDTO dto) {
 
-        ExperienceResponseDTO updatedExperience = experienceService.editById(id, dto);
+        ExperienceResponseDTO updatedExperience = experienceService.editById(user.getId(), id, dto);
 
         return ResponseEntity.ok(WebResponse.<ExperienceResponseDTO>builder()
                 .data(updatedExperience)
@@ -61,9 +83,20 @@ public class ExperienceController {
     @DeleteMapping("/{id}")
     public ResponseEntity<WebResponse<String>> deleteExperienceById(
             @PathVariable("id") Long id,
-            @RequestHeader("Authorization") String token) {
+            @AuthenticationPrincipal User user) {
 
-        experienceService.deleteById(id);
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(WebResponse.<String>builder()
+                    .errors("Unauthorized access")
+                    .build());
+        }
+        try {
+            experienceService.deleteById(id);
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(WebResponse.<String>builder()
+                    .errors(e.getMessage())
+                    .build());
+        }
 
         return ResponseEntity.ok(WebResponse.<String>builder()
                 .data("Experience with id " + id + " deleted")
